requirements.txt
flask==2.3.3
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
werkzeug==3.0.1
streamlit==1.35.0
requests==2.31.0
PyJWT==2.8.0